+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Support EnvArg for whenEnvVar
 - Add help methods to build EnvArg
+- For top level condition of stage or pipeline it should combine all condition with && rule
 
 ## [0.4.3] - 2023-09-03